add java docs and comments
