+++ src/main/java/entity/sophia/GoalInfo.java
@@ -60,4 +60,3 @@ public class GoalInfo {
         this.targetTaskInfo = targetTaskInfo;
     }
 }
-
