--- src/main/java/interface_adapter/Sophia/create_goal/CreatedGoalState.java
@@ -1,13 +1,0 @@
-package interface_adapter.Sophia.create_goal;
-
-public class CreatedGoalState {
-    private String message = "";
-
-    public String getMessage() {
-        return message;
-    }
-
-    public void setMessage(String message) {
-        this.message = message;
-    }
-}
