+++ src/main/java/views/CreateGoalView.java
@@ -3,8 +3,8 @@ package views;
 import entity.Angela.Task.Task;
 import entity.BeginAndDueDates.BeginAndDueDates;
 import entity.info.Info;
-import interface_adapter.Sophia.create_goal.CreatedGoalViewModel;
-import interface_adapter.Sophia.create_goal.CreateGoalController;
+import interface_adapter.sophia.create_goal.CreatedGoalViewModel;
+import interface_adapter.sophia.create_goal.CreateGoalController;
 import entity.Sophia.Goal;
 import javax.swing.*;
 import java.awt.*;
